automatically embed study programs
