+++ embedder-bridge/src/main/kotlin/com/nixops/embedderbridge/scheduler/EmbeddingScheduler.kt
@@ -1,0 +1,16 @@
+package com.nixops.embedderbridge.scheduler
+
+import com.nixops.embedderbridge.services.EmbeddingBridgeService
+import org.springframework.scheduling.annotation.Scheduled
+import org.springframework.stereotype.Component
+
+@Component
+class EmbeddingScheduler(
+    private val embeddingBridgeService: EmbeddingBridgeService,
+    private val fixedRateString: String
+) {
+  @Scheduled(fixedRateString = "#{@fixedRateString}")
+  fun embed() {
+    embeddingBridgeService.embedNextCandidate()
+  }
+}
