+++ embedder-bridge/src/main/resources/application.properties
@@ -3,5 +3,6 @@ spring.application.name=embedder-bridge
 server.port=8040
 
 # app config
-embedder-bridge.scraper.base-url=http://localhost:8080
-embedder-bridge.embedding.base-url=http://localhost:8000
+embedding-bridge.scraper.base-url=http://localhost:8080
+embedding-bridge.embedding.base-url=http://localhost:8000
+embedding-bridge.interval=10s
